vector-search/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
annoy==1.17.3
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
exceptiongroup==1.2.0
Flask==3.0.2
frozenlist==1.4.1
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
jq==1.6.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
langchain-google-vertexai==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.24
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.9.15
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
tenacity==8.2.3
types-protobuf==4.24.0.20240311
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.18.0
